// src/app/Models/almacen/OrdenDescarga/ordenDescarga-model.ts
export class OrdenDescarga {
    IdOrdenDescarga: number;
    Folio: number;
    FechaLlegada: Date;
    IdProveedor: number;
    Proveedor: string;
    IdPedido: number;
    PO: string;
    Fletera: string;
    Caja: string;
    Sacos: string;
    Kg: string;
    Chofer: string;
    Origen: string;
    Destino: string;
    Observaciones: string;
    Estatus: string;
    FechaInicioDescarga: Date;
    FechaFinalDescarga: Date;
    FechaExpedicion: Date;
    IdUsuario: number;
    Usuario: string;
}